Rename `rustc_middle/error.rs` to `diagnostics.rs`
