+++ compiler/rustc_middle/src/diagnostics.rs
