+++ compiler/rustc_middle/src/middle/lang_items.rs
@@ -19,7 +19,8 @@ impl<'tcx> TyCtxt<'tcx> {
     /// If not found, fatally aborts compilation.
     pub fn require_lang_item(self, lang_item: LangItem, span: Span) -> DefId {
         self.lang_items().get(lang_item).unwrap_or_else(|| {
-            self.dcx().emit_fatal(crate::error::RequiresLangItem { span, name: lang_item.name() });
+            self.dcx()
+                .emit_fatal(crate::diagnostics::RequiresLangItem { span, name: lang_item.name() });
         })
     }
