+++ compiler/rustc_middle/src/traits/specialization_graph.rs
@@ -4,7 +4,7 @@ use rustc_hir::def_id::{DefId, DefIdMap};
 use rustc_hir::find_attr;
 use rustc_macros::{StableHash, TyDecodable, TyEncodable};
 
-use crate::error::StrictCoherenceNeedsNegativeCoherence;
+use crate::diagnostics::StrictCoherenceNeedsNegativeCoherence;
 use crate::ty::fast_reject::SimplifiedType;
 use crate::ty::{self, TyCtxt, TypeVisitableExt};
